requirements.txt
arcade==3.3.3
pygame-ce==2.5.7